START: Extract control loop
